Including the "success" function in Failure.
Yes, it's ironic...

Point here is that, although this should always be a synonym to "pure"
and "return", we sometimes do not want to specify the full power of an
ApplicativeFailure or MonadFailure.

Bumped version to 0.2.0. This is necesary also for the previous commit,
since it adds new instances. (In this case, technically the PVP requires
the version bump, but I don't think we should be worried about breaking
code that includes orphan instances.)